\name{mig}
\alias{mig}
\title{Memory-efficient Implementation of Gabriel et al. 2002 haplotype block definition}
\description{
	Function for the efficient whole-genome haplotype block partitioning.
	Haplotype blocks are defined based on D' coefficient of linkage disequilibrium (Gabriel et al., 2002).
}
\usage{mig()}
\author{Daniel Taliun}
\keyword{misc}
\keyword{utilities}
\keyword{package}
\examples{}